fix(qa): convert uPlatformType from int to float for WebGL 1.0 compat
WebGL 1.0 (Three.js ShaderMaterial target) does not reliably support
integer uniforms across all platforms. Convert uPlatformType uniform
from `int` to `float` throughout the pipeline:

- SdfParamRegistry: glslType int→float, clamp range 2-5→2-15
- sdfPlatforms.ts: sdPlatform accepts float, converts to int internally
- backgroundRaymarch.ts: fix platform Y offset mismatch (2.5→4.0) in
  both map() and main() detection to match updated uPlatformZ range
- DebugGUI: use string keys for platform type dropdown (lil-gui compat)
- All 87 tests updated and passing, production build clean

+++ frontend/src/components/tree/scene/SdfParamRegistry.ts
@@ -181,7 +181,7 @@ export const SDF_PARAM_REGISTRY: SdfParamEntry[] = [
   // --- platform (2 entries) ---
   {
     name: 'uPlatformType',
-    glslType: 'int',
+    glslType: 'float',
     tsKey: 'bgPlatformType',
     defaultValue: 0,
     category: 'geometry',
